// src/mcp/types.ts
export type ProblemType = 
  | 'vehicle_routing'
  | 'pickup_delivery'
  | 'job_shop'
  | 'project_scheduling'
  | 'resource_scheduling'
  | 'fleet_scheduling'
  | 'multi_depot_routing'
  | 'bin_packing'
  | 'traveling_salesman'
  | 'assignment'
  | 'flow_shop'
  | 'nurse_scheduling'
  | 'inventory_optimization'
  | 'production_planning'
  | 'custom'; 